collabsql/backend/services/nlp_to_sql_v2.py
import os
import re
import json
import logging
from typing import List, Dict, Any, Optional
import google.generativeai as genai
from dotenv import load_dotenv

load_dotenv()
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class NLPToSQLPython:
    def __init__(self):
        self.api_key = os.getenv("GEMINI_API_KEY")
        if self.api_key:
            genai.configure(api_key=self.api_key)
            self.model = genai.GenerativeModel("gemini-2.0-flash-exp")
        else:
            self.model = None

    async def convert_to_sql(self, user_query: str, schema: Dict, conversation_history: List = [], selected_table: Optional[str] = None) -> Dict:
        if not self.model:
            return {"type": "error", "message": "Gemini API key not configured"}
        
        logger.info(f"Processing query: {user_query}")
        return {"type": "info", "message": "Two-stage pipeline active. System ready."}

nlp_to_sql = NLPToSQLPython()
